backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
PyPDF2==3.0.1
numpy==2.1.0
openai==1.54.0
tiktoken==0.8.0